+++ basex-core/src/main/java/org/basex/query/value/map/XQAtmStrMap.java
@@ -56,13 +56,13 @@ public final class XQAtmStrMap extends XQHashMap {
   }
 
   @Override
-  public Atm keyAt(final int index) {
-    return Atm.get(map.key(index + 1));
+  public Atm keyAt(final long index) {
+    return Atm.get(map.key((int) index + 1));
   }
 
   @Override
-  public Str valueAt(final int index) {
-    return Str.get(map.value(index + 1));
+  public Str valueAt(final long index) {
+    return Str.get(map.value((int) index + 1));
   }
 
   @Override
