+++ basex-core/src/main/java/org/basex/query/value/map/XQIntMap.java
@@ -60,13 +60,13 @@ public final class XQIntMap extends XQHashMap {
   }
 
   @Override
-  public Itr keyAt(final int index) {
-    return Itr.get(map.key(index + 1));
+  public Itr keyAt(final long index) {
+    return Itr.get(map.key((int) index + 1));
   }
 
   @Override
-  public Itr valueAt(final int index) {
-    return Itr.get(map.value(index + 1));
+  public Itr valueAt(final long index) {
+    return Itr.get(map.value((int) index + 1));
   }
 
   @Override
