+++ basex-core/src/main/java/org/basex/query/value/map/XQStrMap.java
@@ -56,13 +56,13 @@ public final class XQStrMap extends XQHashMap {
   }
 
   @Override
-  public Str keyAt(final int index) {
-    return Str.get(map.key(index + 1));
+  public Str keyAt(final long index) {
+    return Str.get(map.key((int) index + 1));
   }
 
   @Override
-  public Str valueAt(final int index) {
-    return Str.get(map.value(index + 1));
+  public Str valueAt(final long index) {
+    return Str.get(map.value((int) index + 1));
   }
 
   @Override
